homework_2/pyproject.toml
[project]
name = "homework-2"
version = "0.1.0"
description = ""
authors = [
    {name = "liskunovich",email = "devid02ru@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "setuptools (>=77.0.3,<78.0.0)",
    "wheel (>=0.45.1,<0.46.0)",
    "twine (>=6.1.0,<7.0.0)",
    "latex-table-generator-homework-itmo-liskunovich (>=0.3.1,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
